src/requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
sentence_transformers
langchain-community
tiktoken

